Beginning repo cleanup
