+++ giga-api/0_images/grayscale/grayscale_main.cpp
