+++ giga-api/0_images/sharpening/sharpening_main.cpp
