+++ giga-api/0_images/upsampling/upsampled_image.png
