+++ giga-api/1_basic/fft/input_signal.txt
